Replaces z3c.saconfig with a custom session manager.
Until now seantis.reservation didn't really work well with concurrent requests. One user allocating space could conflict with another one doing the same thing, if the database transactions were run at the same time.

This commit introduces a session manager which offers two kinds of sessions. Serial sessions (critical) and main sessions (for reading). The serial session ensures that all transactions used on it are processed one at the time by postgres.

For this to work the session makes use of true serialization added in postgres 9.1. Using this mechanism functions can be decorated as 'serialized'. Everything within their scope will then be handled by true serialization. That is if two transactions work with the same data in any way, postgres will block conflicts. They can then be repeated (which is not done yet), or the user can be informed that the data he intents to work with was changed at the very moment he decided to change them.

A couple of important things

 * use seantis.reservation.Session for all your sessions
 * use postgres 9.1 or newer (this dependancy will be mandatory in the future)
 * tests need to be rewritten to some extent as they do only work with sqlite
 * transaction should be as lean as possible as true serialization expensive